+++ MOD开发/WhereIsMyAmulet/WhereIsMyAmulet/Properties/AssemblyInfo.cs
@@ -1,0 +1,10 @@
+using System.Reflection;
+using System.Runtime.InteropServices;
+
+[assembly: AssemblyTitle("WhereIsMyAmulet")]
+[assembly: AssemblyDescription("Shows the distance to dropped amulet fragments in PEAK.")]
+[assembly: AssemblyProduct("WhereIsMyAmulet")]
+[assembly: ComVisible(false)]
+[assembly: Guid("ce0f5e2c-08a1-4f5f-8da4-a91c5c43a5a7")]
+[assembly: AssemblyVersion("1.0.0.0")]
+[assembly: AssemblyFileVersion("1.0.0.0")]